requirements.txt
Backend:

Flask
mysql-connector-python
Pillow
requests
flask-caching
python-dotenv

Frontend:

nicegui
httpx